poc/feishu/requirements.txt
lark-oapi>=1.3.0
flask>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
